wpt_babel/content-security-policy/script-src/nonce-enforce-blocked.html

<!DOCTYPE html>
<meta content="script-src 'nonce-abc'" http-equiv="Content-Security-Policy"/>
<script nonce="abc" src="/resources/testharness.js">
</script>
<script nonce="abc" src="/resources/testharnessreport.js">
</script>
<script nonce="abc">
 "use strict";

var t = async_test("Unnonced scripts generate reports.");
var events = 0;
var firstLine = 43;
var expectations = {};
expectations[firstLine] = true;
expectations[firstLine + 3] = true;
expectations[firstLine + 6] = true;
expectations[firstLine + 9] = true;
expectations[firstLine + 12] = true;
expectations[firstLine + 15] = true;
expectations[firstLine + 18] = true;
expectations[firstLine + 21] = true;
expectations[firstLine + 24] = true;
expectations[firstLine + 28] = true;
expectations["/content-security-policy/support/nonce-should-be-blocked.js?1"] = true;
expectations["/content-security-policy/support/nonce-should-be-blocked.js?2"] = true;
expectations["/content-security-policy/support/nonce-should-be-blocked.js?3"] = true;
expectations["/content-security-policy/support/nonce-should-be-blocked.js?4"] = true;
expectations["/content-security-policy/support/nonce-should-be-blocked.js?5"] = true;
expectations["/content-security-policy/support/nonce-should-be-blocked.js?6"] = true;
expectations["/content-security-policy/support/nonce-should-be-blocked.js?7"] = true;
document.addEventListener('securitypolicyviolation', t.step_func(function (e) {
  if (e.lineNumber) {
    // Verify that the line is expected, then clear the expectation:
    assert_true(expectations[e.lineNumber], "Line number: " + e.lineNumber);
    assert_equals(e.blockedURI, "inline");
  } else {
    // Otherwise, verify that the URL is expected, then clear the expectation:
    var url = new URL(e.blockedURI);
    assert_true(expectations[url.pathname + url.search], "URL: " + e.blockedURI);
  }
  events++;
  if (events == Object.keys(expectations).length) t.done();
}));
</script>
<script>
 "use strict";

t.unreached_func("No nonce, no execution.")();
</script>
<script nonce="xyz">
 "use strict";

t.unreached_func("Bad nonce, no execution.")();
</script>
<script <script="" nonce="abc">
 "use strict";

t.unreached_func("'<script' attribute, no execution.")();
</script>
<script attribute<script="" nonce="abc">
 "use strict";

t.unreached_func("'attribute<script', no execution.")();
</script>
<script attribute<style="value" nonce="abc">
 "use strict";

t.unreached_func("'attribute<style' attribute, no execution.")();
</script>
<script attribute="&lt;script" nonce="abc">
 "use strict";

t.unreached_func("'<script' value, no execution.")();
</script>
<script attribute="value&lt;script" nonce="abc">
 "use strict";

t.unreached_func("'value<script', no execution.")();
</script>
<script attribute="" nonce="abc">
 "use strict";

t.unreached_func("Duplicate attribute, no execution.")();
</script>
<script attribute="&lt;style" nonce="abc">
 "use strict";

t.unreached_func("2# Duplicate attribute, no execution.")();
</script>
<svg xmlns="http://www.w3.org/2000/svg">
 <script attribute="" nonce="abc">
  "use strict";

t.unreached_func("Duplicate attribute in SVG, no execution.")();
 </script>
</svg>
<script <script="" nonce="abc" src="../support/nonce-should-be-blocked.js?1">
</script>
<script attribute="&lt;script" nonce="abc" src="../support/nonce-should-be-blocked.js?2">
</script>
<script <style="" nonce="abc" src="../support/nonce-should-be-blocked.js?3">
</script>
<script attribute="&lt;style" nonce="abc" src="../support/nonce-should-be-blocked.js?4">
</script>
<script attribute="" nonce="abc" src="../support/nonce-should-be-blocked.js?5">
</script>
<script attribute<script="" nonce="abc" src="../support/nonce-should-be-blocked.js?6">
</script>
<script attribute="value&lt;script" nonce="abc" src="../support/nonce-should-be-blocked.js?7">
</script>
